+++ src/modules/app/config/profiles/edc-ui-profile.ts
@@ -1,6 +1,5 @@
 import {EDC_UI_PROFILE_DATA} from './edc-ui-profile-data';
 
-
 /**
  * Available Configuration Profiles.
  */