+++ src/modules/edc-demo/components/policy-rule-viewer/policy-rule-viewer.component.ts
@@ -1,4 +1,4 @@
-import {Component, Input, OnInit} from '@angular/core';
+import {Component, Input} from '@angular/core';
 import {Rule} from '../../../edc-dmgmt-client/model/rule';
 
 @Component({
@@ -6,11 +6,7 @@ import {Rule} from '../../../edc-dmgmt-client/model/rule';
   templateUrl: './policy-rule-viewer.component.html',
   styleUrls: ['./policy-rule-viewer.component.scss'],
 })
-export class PolicyRuleViewerComponent implements OnInit {
+export class PolicyRuleViewerComponent {
   @Input() rules: Rule[] | undefined = [];
   @Input() title: string = '';
-
-  constructor() {}
-
-  ngOnInit(): void {}
 }