+++ src/modules/edc-demo/models/asset.ts
@@ -10,7 +10,7 @@ import {TransportModeSelectItem} from '../components/transport-mode-select/trans
  */
 export interface Asset {
   id: string;
-  name: string | null;
+  name: string;
   version: string | null;
   contentType: string | null;
   originator: string | null;
@@ -29,29 +29,12 @@ export interface Asset {
   dataModel: string | null;
   geoReferenceMethod: string | null;
   transportMode: TransportModeSelectItem | null;
+
+  // Unhandled Additional Properties
+  additionalProperties: AdditionalAssetProperty[];
 }
 
-export function dummyAsset(id: string): Asset {
-  return {
-    id,
-    name: id,
-    version: null,
-    contentType: null,
-    originator: null,
-    originatorOrganization: null,
-    keywords: [],
-
-    description: null,
-    language: null,
-    publisher: null,
-    standardLicense: null,
-    endpointDocumentation: null,
-
-    // MDS Specific
-    dataCategory: null,
-    dataSubcategory: null,
-    dataModel: null,
-    geoReferenceMethod: null,
-    transportMode: null,
-  };
+export interface AdditionalAssetProperty {
+  key: string;
+  value: string;
 }