+++ src/modules/edc-dmgmt-client/model/asset.ts
@@ -12,6 +12,5 @@
 
 
 export interface Asset {
-    properties?: { [key: string]: string; };
+  properties?: { [key: string]: string; };
 }
-